+++ dev/install-common-deps.sh
@@ -30,12 +30,22 @@ source activate test-environment
 python --version
 pip install --upgrade pip==19.3.1
 
+if [[ "$MLFLOW_SKINNY" == "true" ]]; then
+  pip install . --upgrade
+else
+  pip install .[extras] --upgrade
+fi
+export MLFLOW_HOME=$(pwd)
+
 # Install Python test dependencies only if we're running Python tests
 if [[ "$INSTALL_SMALL_PYTHON_DEPS" == "true" ]]; then
   # When downloading large packages from PyPI, the connection is sometimes aborted by the
   # remote host. See https://github.com/pypa/pip/issues/8510.
   # As a workaround, we retry installation of large packages.
   retry-with-backoff pip install --quiet -r ./dev/small-requirements.txt
+fi
+if [[ "$INSTALL_SKINNY_PYTHON_DEPS" == "true" ]]; then
+  retry-with-backoff pip install --quiet -r ./dev/skinny-requirements.txt
 fi
 if [[ "$INSTALL_LARGE_PYTHON_DEPS" == "true" ]]; then
   retry-with-backoff pip install --quiet -r ./dev/large-requirements.txt
@@ -47,9 +57,6 @@ if [[ "$INSTALL_LARGE_PYTHON_DEPS" == "true" ]]; then
   chmod 777 $(find $CONDA_DIR/envs/test-environment/ -path "*bin/spark-*")
   ls -lha $(find $CONDA_DIR/envs/test-environment/ -path "*bin/spark-*")
 fi
-
-pip install .[extras]
-export MLFLOW_HOME=$(pwd)
 
 # Print current environment info
 pip list
