+++ tests/projects/test_projects_cli.py
@@ -187,3 +187,29 @@ def test_run_databricks_cluster_spec(tmpdir):
             env={"MLFLOW_TRACKING_URI": "databricks://profile"},
         )
         assert res.exit_code != 0
+
+
+def test_mlflow_run():
+    with mock.patch("mlflow.cli.projects") as mock_projects:
+        result = CliRunner().invoke(cli.run)
+        mock_projects.run.assert_not_called()
+        assert "Missing argument 'URI'" in result.output
+
+    with mock.patch("mlflow.cli.projects") as mock_projects:
+        CliRunner().invoke(cli.run, ["project_uri"])
+        mock_projects.run.assert_called_once()
+
+    with mock.patch("mlflow.cli.projects") as mock_projects:
+        CliRunner().invoke(cli.run, ["--experiment-id", "5", "project_uri"])
+        mock_projects.run.assert_called_once()
+
+    with mock.patch("mlflow.cli.projects") as mock_projects:
+        CliRunner().invoke(cli.run, ["--experiment-name", "random name", "project_uri"])
+        mock_projects.run.assert_called_once()
+
+    with mock.patch("mlflow.cli.projects") as mock_projects:
+        result = CliRunner().invoke(
+            cli.run, ["--experiment-id", "51", "--experiment-name", "name blah", "uri"]
+        )
+        mock_projects.run.assert_not_called()
+        assert "Specify only one of 'experiment-name' or 'experiment-id' options." in result.output
